refactor(test): dedicated e2e TestContext module
Moves the router's TestContext to its own file/module.
